# services/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ronald-GI API
  description: |
    Local-first autonomous research assistant API.

    Ronald-GI learns from your browsing patterns, monitors sources you care about,
    and proactively surfaces insights tailored to your interests.

    ## Key Features
    - **Personalized Learning**: System learns from your interactions
    - **Proactive Briefings**: Daily summaries of what matters
    - **Concept Tracking**: Monitors topics you care about
    - **Research Automation**: Autonomous research on your behalf

    ## Authentication
    This API is designed for local use and does not require authentication.
    In production deployments, add appropriate authentication middleware.
  version: 1.0.0
  contact:
    name: Ronald-GI
  license:
    name: MIT
servers:
  - url: http://localhost:3100/api
    description: Local development server
  - url: http://localhost:3100/api
    description: Docker container
...
paths:
...
  /briefing/alerts/{id}/dismiss:
    post:
      summary: Dismiss an alert
      tags: [Briefing]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert dismissed
...
  /jobs:
    get:
      summary: List jobs
      description: Get background jobs
      tags: [Jobs]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
  /jobs/{id}:
    get:
      summary: Get job status
      tags: [Jobs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'

  # ========== Visits ==========
...
components:
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: string
        job_type:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        priority:
          type: integer
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...